🦈 IMP: Saturating Gravitational History Heuristic Formula
--- STC ---
Elo   | 14.42 +- 7.10 (95%)
SPRT  | 10.0+0.10s Threads=1 Hash=16MB
LLR   | 2.96 (-2.94, 2.94) [0.00, 5.00]
Games | N: 3374 W: 1074 L: 934 D: 1366
Penta | [52, 331, 793, 447, 64]

--- LTC ---
Elo   | 15.78 +- 7.13 (95%)
SPRT  | 60.0+0.60s Threads=1 Hash=128MB
LLR   | 2.96 (-2.94, 2.94) [0.00, 5.00]
Games | N: 2732 W: 800 L: 676 D: 1256
Penta | [20, 260, 692, 364, 30]

Bench: 5523803

+++ src/Engine/Move/HistoryTable.h
@@ -14,7 +14,7 @@
 namespace StockDory
 {
 
-    using HistoryTable = std::array<std::array<std::array<int32_t, 64>, 6>, 2>;
+    using HistoryTable = std::array<std::array<std::array<int16_t, 64>, 6>, 2>;
 
 } // StockDory
